requirements.txt
aioboto3
quart
